Split list handers into individual files
