+++ internal/pkg/keybindings/list_back.go
@@ -1,0 +1,26 @@
+package keybindings
+
+import (
+	"github.com/awesome-gocui/gocui"
+	"github.com/lawrencegripper/azbrowse/internal/pkg/views"
+)
+
+type ListBackHandler struct {
+	ListHandler
+	List *views.ListWidget
+}
+
+func NewListBackHandler(list *views.ListWidget) *ListBackHandler {
+	handler := &ListBackHandler{
+		List: list,
+	}
+	handler.id = HandlerIDListBack
+	return handler
+}
+
+func (h ListBackHandler) Fn() func(g *gocui.Gui, v *gocui.View) error {
+	return func(g *gocui.Gui, v *gocui.View) error {
+		h.List.GoBack()
+		return nil
+	}
+}
