+++ internal/pkg/keybindings/list_back_legacy.go
@@ -1,0 +1,26 @@
+package keybindings
+
+import (
+	"github.com/awesome-gocui/gocui"
+	"github.com/lawrencegripper/azbrowse/internal/pkg/views"
+)
+
+type ListBackLegacyHandler struct {
+	ListHandler
+	List *views.ListWidget
+}
+
+func NewListBackLegacyHandler(list *views.ListWidget) *ListBackLegacyHandler {
+	handler := &ListBackLegacyHandler{
+		List: list,
+	}
+	handler.id = HandlerIDListBackLegacy
+	return handler
+}
+
+func (h ListBackLegacyHandler) Fn() func(g *gocui.Gui, v *gocui.View) error {
+	return func(g *gocui.Gui, v *gocui.View) error {
+		h.List.GoBack()
+		return nil
+	}
+}
