+++ internal/pkg/keybindings/list_down.go
@@ -1,0 +1,26 @@
+package keybindings
+
+import (
+	"github.com/awesome-gocui/gocui"
+	"github.com/lawrencegripper/azbrowse/internal/pkg/views"
+)
+
+type ListDownHandler struct {
+	ListHandler
+	List *views.ListWidget
+}
+
+func NewListDownHandler(list *views.ListWidget) *ListDownHandler {
+	handler := &ListDownHandler{
+		List: list,
+	}
+	handler.id = HandlerIDListDown
+	return handler
+}
+
+func (h ListDownHandler) Fn() func(g *gocui.Gui, v *gocui.View) error {
+	return func(g *gocui.Gui, v *gocui.View) error {
+		h.List.MoveDown()
+		return nil
+	}
+}
