+++ internal/pkg/keybindings/list_expand.go
@@ -1,0 +1,26 @@
+package keybindings
+
+import (
+	"github.com/awesome-gocui/gocui"
+	"github.com/lawrencegripper/azbrowse/internal/pkg/views"
+)
+
+type ListExpandHandler struct {
+	ListHandler
+	List *views.ListWidget
+}
+
+func NewListExpandHandler(list *views.ListWidget) *ListExpandHandler {
+	handler := &ListExpandHandler{
+		List: list,
+	}
+	handler.id = HandlerIDListExpand
+	return handler
+}
+
+func (h ListExpandHandler) Fn() func(g *gocui.Gui, v *gocui.View) error {
+	return func(g *gocui.Gui, v *gocui.View) error {
+		h.List.ExpandCurrentSelection()
+		return nil
+	}
+}
