+++ internal/pkg/keybindings/list_refresh.go
@@ -1,0 +1,26 @@
+package keybindings
+
+import (
+	"github.com/awesome-gocui/gocui"
+	"github.com/lawrencegripper/azbrowse/internal/pkg/views"
+)
+
+type ListRefreshHandler struct {
+	ListHandler
+	List *views.ListWidget
+}
+
+func NewListRefreshHandler(list *views.ListWidget) *ListRefreshHandler {
+	handler := &ListRefreshHandler{
+		List: list,
+	}
+	handler.id = HandlerIDListRefresh
+	return handler
+}
+
+func (h ListRefreshHandler) Fn() func(g *gocui.Gui, v *gocui.View) error {
+	return func(g *gocui.Gui, v *gocui.View) error {
+		h.List.Refresh()
+		return nil
+	}
+}
